+++ src/game/continuous_effects.rs
@@ -307,6 +307,7 @@ impl Game {
             | EffectDef::EmptyManaPool { .. }
             | EffectDef::LoseLife { .. }
             | EffectDef::LoseTheGame { .. }
+            | EffectDef::Regenerate { .. }
             | EffectDef::Tap { .. }
             | EffectDef::Untap { .. }
             | EffectDef::PreventCombatDamageThisTurn { .. }